### 2025-11-06 #9362 @editorialbot
[PRE REVIEW]: ParallelKDE.jl: A Package for Highly Parallel Kernel Density Estimation
**Submitting author:** @Clownshift (<a href="http://orcid.org/0000-0002-6227-5443">Patrick K. Quoika</a>)
**Repository:** https://github.com/chrissm23/ParallelKDE.jl.git
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
