### 2025-11-06 3598853
fix references
fixed a few dois, but I am still waiting for the preprint of the other
article to actually receive a doi. This will probably take another 48
hrs.
