+++ paper/paper.bib
@@ -25,7 +25,6 @@
 	publisher = {Chapman \& Hall},
 	author = {Wand, Matt P. and Jones, M. C.},
 	year = {1995},
-	file = {Table of Contents PDF:/home/ch_sustay/Zotero/storage/YJQ6IULW/Wand and Jones - 1995 - Kernel smoothing.pdf:application/pdf},
 }
 
 @article{kraml_x-entropy_2021,
@@ -44,12 +43,10 @@
 	month = apr,
 	year = {2021},
 	pages = {1533--1538},
-	file = {Full Text PDF:/home/ch_sustay/Zotero/storage/4C2IXMJN/Kraml et al. - 2021 - X-Entropy A Parallelized Kernel Density Estimator with Automated Bandwidth Selection to Calculate E.pdf:application/pdf},
 }
 
 @article{sustay_martinez_novel_2025,
 	title = {Novel {Rapid} {Approach} for {Adaptive} {Gaussian} {Kernel} {Density} {Estimation}: {Gridpoint}-wise {Propagation} of {Anisotropic} {Diffusion} {Equation}},
-	issn = {XXX-XXX, XXXX-XXXX},
 	doi = {10.1007/s41060-XXX-XXXX-X},
 	language = {English},
 	journal = {International Journal of Data Science and Analytics},
@@ -85,7 +82,6 @@
 	url = {https://zenodo.org/record/2392268},
 	abstract = {Kernel Density Estimation in Python.},
 	urldate = {2025-10-09},
-	publisher = {Zenodo},
 	author = {Odland, Tommy},
 	month = dec,
 	year = {2018},
@@ -99,6 +95,7 @@
 	booktitle = {Proceedings of the 9th {Python} in {Science} {Conference}},
 	author = {Seabold, Skipper and Perktold, Josef},
 	year = {2010},
+	doi = {10.25080/majora-92bf1922-011},
 }
 
 @article{pedregosa_scikit-learn_2011,
@@ -129,7 +126,6 @@
 	month = mar,
 	year = {2020},
 	pages = {261--272},
-	file = {Full Text:/home/ch_sustay/Zotero/storage/8X4HWUB8/Virtanen et al. - 2020 - SciPy 1.0 fundamental algorithms for scientific computing in Python.pdf:application/pdf},
 }
 
 @article{jones_brief_1996,
@@ -168,6 +164,7 @@
 	author = {Scott, David W.},
 	month = aug,
 	year = {1992},
+	doi = {10.1002/9780470316849},
 }
 
 @article{bezanson_julia_2017,
@@ -184,7 +181,6 @@
 	month = jan,
 	year = {2017},
 	pages = {65--98},
-	file = {Submitted Version:/home/ch_sustay/Zotero/storage/T3AQHLNM/Bezanson et al. - 2017 - Julia A Fresh Approach to Numerical Computing.pdf:application/pdf},
 }
 
 @article{besard_effective_2019,
@@ -200,7 +196,6 @@
 	month = apr,
 	year = {2019},
 	pages = {827--841},
-	file = {Submitted Version:/home/ch_sustay/Zotero/storage/EY2BLCBD/Besard et al. - 2019 - Effective Extensible Programming Unleashing Julia on GPUs.pdf:application/pdf},
 }
 
 @article{besard_rapid_2019,
@@ -215,7 +210,6 @@
 	month = jun,
 	year = {2019},
 	pages = {29--46},
-	file = {Accepted Version:/home/ch_sustay/Zotero/storage/WQQCHL6R/Besard et al. - 2019 - Rapid software prototyping for heterogeneous and distributed platforms.pdf:application/pdf},
 }
 
 @misc{rowley_pythoncalljl_2022,
