+++ paper/paper.md
@@ -18,7 +18,8 @@ authors:
     corresponding: true
     affiliation: 1
 affiliations:
-  - name: Center for Functional Protein Assemblies, Technical University of Munich, Ernst-Otto-Fischer-Str. 8, Garching 85748, Germany
+  - index: 1
+  - name: Center for Functional Protein Assemblies, School of Natural Sciences, Technical University of Munich, Ernst-Otto-Fischer-Str. 8, Garching 85748, Germany
 date: 05 November 2025
 bibliography: paper.bib
 ---
