### 2025-11-06 50d5457
Fix list in paper.md affiliations section
