+++ paper/paper.bib
@@ -47,9 +47,9 @@
 
 @article{sustay_martinez_novel_2025,
 	title = {Novel {Rapid} {Approach} for {Adaptive} {Gaussian} {Kernel} {Density} {Estimation}: {Gridpoint}-wise {Propagation} of {Anisotropic} {Diffusion} {Equation}},
-	doi = {10.1007/s41060-XXX-XXXX-X},
+	doi = {10.21203/rs.3.rs-8038701/v1},
 	language = {English},
-	journal = {International Journal of Data Science and Analytics},
+	journal = {Researchsquare},
 	author = {Sustay Martinez, Christian and Quoika, Patrick K. and Zacharias, Martin},
 	year = {2025},
 }
