+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api_reference/kdes/#APIKDEs","page":"KDE Interface","title":"KDE Interface","text":"","category":"section"},{"location":"api_reference/kdes/","page":"KDE Interface","title":"KDE Interface","text":"KDE objects store the final result of the estimation. This usually consists of an array of densities mapped to the selected grid. They also store the samples used for the estimation.","category":"page"},{"location":"api_reference/kdes/","page":"KDE Interface","title":"KDE Interface","text":"There are currently two concrete KDE types, one for CPU and one for CUDA devices. However, the interface is otherwise the same, so that it is possible to use the same code for both devices. It is also possible to create custom KDE objects that conform to the interface.","category":"page"},{"location":"api_reference/kdes/#ParallelKDE.KDEs.AbstractKDE","page":"KDE Interface","title":"ParallelKDE.KDEs.AbstractKDE","text":"AbstractKDE{N,T,S}\n\nSupertype for kernel density estimation (KDEs) with N dimensions, T type for density values, and S type for data points.\n\n\n\n\n\n","category":"type"},{"location":"api_reference/kdes/#ParallelKDE.KDEs.KDE","page":"KDE Interface","title":"ParallelKDE.KDEs.KDE","text":"KDE{N,T<:Real,S<:Real}\n\nCPU object for kernel density estimation (KDE) with N dimensions, T type for density values, and S type for data points.\n\n\n\n\n\n","category":"type"},{"location":"api_reference/kdes/#ParallelKDE.KDEs.CuKDE","page":"KDE Interface","title":"ParallelKDE.KDEs.CuKDE","text":"CuKDE{N,T<:Real,S<:Real}\n\nCUDA object for kernel density estimation (KDE) with N dimensions, T type for density values, and\n\n\n\n\n\n","category":"type"},{"location":"api_reference/kdes/","page":"KDE Interface","title":"KDE Interface","text":"The following is a set of functions that extract information from KDE objects.","category":"page"},{"location":"api_reference/kdes/#ParallelKDE.Devices.get_device","page":"KDE Interface","title":"ParallelKDE.Devices.get_device","text":"get_device(device::Any)\n\nObtain the device object for a given device type.\n\nIf the method is called with an unsupported type, it returns a DeviceNotSpecified object.\n\n\n\n\n\nget_device(grid::AbstractGrid)\n\nIdentify the device type of the grid, returning IsCPU for CPU grids and IsCUDA for CUDA grids.\n\n\n\n\n\nget_device(kde::AbstractKDE)\n\nIdentify the device type used by the kernel density estimation (KDE) object.\n\n\n\n\n\n","category":"function"},{"location":"api_reference/kdes/#ParallelKDE.KDEs.get_data","page":"KDE Interface","title":"ParallelKDE.KDEs.get_data","text":"get_data(kde::AbstractKDE)\n\nReturn a view of the data stored in the kernel density estimation (KDE) object.\n\n\n\n\n\n","category":"function"},{"location":"api_reference/kdes/#ParallelKDE.KDEs.get_density-Tuple{ParallelKDE.KDEs.AbstractKDE}","page":"KDE Interface","title":"ParallelKDE.KDEs.get_density","text":"get_density(kde::AbstractKDE)\n\nReturn the density values stored in the kernel density estimation (KDE) object.\n\n\n\n\n\n","category":"method"},{"location":"api_reference/kdes/#ParallelKDE.KDEs.get_nsamples","page":"KDE Interface","title":"ParallelKDE.KDEs.get_nsamples","text":"get_nsamples(kde::AbstractKDE)\n\nReturn the number of samples in the kernel density estim
[…]
