### 2025-12-12 #9362 @mantzaris
> 👋 [@mantzaris](https://github.com/mantzaris) - Would you be willing to review this submission to JOSS?
>

Yes, I accept
