### 2025-12-12 #9362 @mantzaris
There is a an issue, I put it on the repo too:
Pkg.add("ParallelKDE") should succeed on Julia 1.10 (or the package should clearly declare it does not support 1.10 and provide guidance).

```
(@v1.10) pkg> add ParallelKDE
    Updating registry at `~/.julia/registries/General.toml`
   Resolving package versions...
ERROR: Unsatisfiable requirements detected for package ParallelKDE [251f600b]:
 ParallelKDE [251f600b] log:
 ├─possible versions are: 1.0.0 or uninstalled
 ├─restricted to versions * by an explicit requirement, leaving only versions: 1.0.0
 └─restricted by julia compatibility requirements to versions: uninstalled — no versions left
```
