+++ .github/workflows/draft-pdf.yml
@@ -30,7 +30,7 @@ jobs:
           journal: joss
           paper-path: paper/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v5
+        uses: actions/upload-artifact@v6
         with:
           name: paper
           path: paper/paper.pdf
