### 2025-12-15 4f8f5c2
included julia compatibility to docs
