### 2025-12-15 e8d9ed6
included julia compatibility to docs
