### 2025-12-17 ab6527b
relaxed julia version to accepct v1.10
