+++ README.md
@@ -32,8 +32,16 @@ or, similarly,
 pkg> add ParallelKDE
 ```
 
+For the latest changes, the package can also be installed directly from the repository with:
+
+```julia
+using Pkg
+Pkg.add(url="https://github.com/chrissm23/ParallelKDE.jl", rev="dev")
+```
+
 > [!IMPORTANT]
-> This package supports Julia v1.11+ and is expected to work on Julia 1.x (SemVer: ^1.11).
+> `ParallelKDE.jl` supports Julia v1.11+ and is expected to work on Julia 1.x (SemVer: ^1.11).
+> In order to use it with Julia v1.10, please install directly from the repository as shown above.
 
 ## Quick Start
 
