+++ Project.toml
@@ -22,7 +22,7 @@ StaticArrays = "^1"
 Statistics = "^1"
 StatsBase = "^0.34"
 Test = "^1"
-julia = "^1.11"
+julia = "^1.10"
 
 [extras]
 Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
