### 2025-12-17 afc8a4c
relaxed julia version to accepct v1.10
