+++ .github/workflows/CI.yml
@@ -23,7 +23,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        julia-version: ["lts", "1"]
+        version: ["lts", "1"]
         os: [ubuntu-latest, windows-latest]
         arch:
           - x64
