### 2025-12-19 #9616 @jbytecode
@hisacro - Edits don't trigger the editorialbot so we need to re-post commands.

You can copy and paste

`@editorialbot generate my checklist`

one more time to generate your check list.
