+++ rot_vs_gradepro/rot_vs_gradepro.jl
@@ -1,0 +1,1676 @@
+### A Pluto.jl notebook ###
+# v0.20.21
+
+using Markdown
+using InteractiveUtils
+
+# ╔═╡ e3d4f7b2-df52-11f0-8ee8-43e8d00d8292
+begin
+	using Plots
+
+	using Distributions
+	using ParallelKDE
+end
+
+# ╔═╡ 9ddbbbca-c6a6-4edf-afb9-a8d299f6c91c
+begin
+	distro = MixtureModel(
+		Normal[
+			Normal(-6.0, 0.2),
+			Normal(0.0, 2)
+		]
+	)
+
+	xs = range(-8, 8, length=250)
+	distro_pdf = pdf.(distro, xs)
+
+	samples = rand(distro, 1, 10000)
+end;
+
+# ╔═╡ 986b6810-3c90-4e4c-8733-45e0c0fc0fc4
+begin
+	grid_bimodal = initialize_grid([xs], device=:cpu)
+	estimation_gradepro = initialize_estimation(
+		samples, grid=grid_bimodal, device=:cpu
+	)
+	estimation_rot = initialize_estimation(
+		samples, grid=grid_bimodal, device=:cpu
+	)
+end;
+
+# ╔═╡ 6daf5cce-1865-4296-93e4-5cc788cf627d
+begin
+	estimate_density!(estimation_gradepro, :gradepro)
+	density_gradepro = get_density(estimation_gradepro)
+end;
+
+# ╔═╡ 17e66106-bc07-42ed-abee-dbc15a5fbb99
+begin
+	estimate_density!(estimation_rot, :rot)
+	density_rot = get_density(estimation_rot)
+end;
+
+# ╔═╡ 7346e2f7-a19d-4504-81e0-ca432eb0a9fc
+begin
+	p = plot(xs, distro_pdf, lw=3, ls=:dash, label="True Distribution")
+	plot!(p, xs, density_gradepro, lw=2, label="GradePro Estimation")
+	plot!(p, xs, density_rot, lw=2, label="Rule of Thumb Estimation")
+	plot!(p, xlabel="Random Variable", ylabel="Density")
+end
+
+# ╔═╡ 00000000-0000-0000-0000-000000000001
+PLUTO_PROJECT_TOML_CONTENTS = """
+[deps]
+Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+ParallelKDE = "251f600b-d60c-4092-b1b9-465b74160c73"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+
+[compat]
+Distributions = "~0.25.122"
+ParallelKDE = "~1.0.1"
+Plots = "~1.41.3"
+"""
+
+# ╔═╡ 00000000-0000-0000-0000-000000000002
+PLUTO_MANIFEST_TOML_CONTENTS = """
+# This file is machine-generated - editing it directly is not advised
+
+julia_version = "1.12.3"
+manifest_format = "2.0"
+project_hash = "e5ab143778342912cef2207972c009d53a520d3c"
+
+[[deps.AbstractFFTs]]
+deps = ["LinearAlgebra"]
+git-tree-sha1 = "d92ad398961a3ed262d8bf04a1a2b8340f915fef"
+uuid = "621f4979-c628-5d54-868e-fcf4e3e8185c"
+version = "1.5.0"
+
+    [deps.AbstractFFTs.extensions]
[…]
