+++ docs/src/contributing.md
@@ -1,8 +1,13 @@
-# [Development](@id Development)
+# [Contributing and Support](@id ContributingSupport)
 
-This section is for contributors and advanced users interested in extending or modifying the package. To request features or report issues, please [open an issue on GitHub](https://github.com/chrissm23/ParallelKDE.jl/issues).
+## Community Guidelines
 
-## Adding new estimators
+- **Support:** For usage questions or help troubleshooting, please [open a GitHub issue](https://github.com/chrissm23/ParallelKDE.jl/issues) including a minimal reproducible example and your Julia and OS versions.
+- **Report issues:** Please [open a GitHub issue](https://github.com/chrissm23/ParallelKDE.jl/issues) and include expected vs. actual behavior, steps to reproduce, and a small input example.
+- **Contribute:** For small changes, feel free to open a pull request directly on [GitHub](https://github.com/chrissm23/ParallelKDE.jl). For larger changes, please [open an issue](https://github.com/chrissm23/ParallelKDE.jl/issues) first to discuss the design, then submit a pull request.
+
+## Adding New Estimators
+
 To add a new estimator:
 
 1. Create a new subtype of the `AbstractEstimator` [interface](@ref "APIEstimators").
@@ -12,7 +17,7 @@ To add a new estimator:
 
 See the [Estimators API](@ref "APIEstimators") for more details on the required methods and how to implement them.
 
-## Current tools
+## Current Tools
 
 The package already provides a number of reusable utilities for building new estimators, including:
 
