+++ docs/Project.toml
@@ -1,5 +1,8 @@
 [deps]
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
+LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
+NPZ = "15e1cf62-19b3-5cfa-8e77-841668bca605"
 ParallelKDE = "251f600b-d60c-4092-b1b9-465b74160c73"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
