+++ docs/src/index.md
@@ -15,6 +15,8 @@ Typical usage involves:
 - Estimating the density with a chosen estimator,
 - Accessing the resulting density.
 
+## Basic Example
+
 For example, to estimate a density on a CPU with a default grid using the [GradePro Estimator](@ref "GradeproEstimator"), you can use:
 
 ```@example 1
@@ -24,7 +26,7 @@ data = randn(1, 10000) # 1-dimensional sample of 10000 points
 
 density_estimation = initialize_estimation(
   data,
-  grid=true,
+  grid=true, # default grid
   device=:cpu,
 )
 estimate_density!(
@@ -50,7 +52,8 @@ which would yield a plot like this:
 
 ```@example 1
 ENV["GKSwstype"]="nul" # hide
-using Plots # hide
+using Plots
+
 p = plot(grid_coordinates, density_true, label="True Density", color=:cornflowerblue, lw=2)
 plot!(p, grid_coordinates, density_estimated, label="Estimated Density", color=:firebrick, lw=2)
 plot!(p, xlabel="Random Variable", ylabel="Density")
@@ -59,7 +62,124 @@ savefig("basic_usage.svg"); nothing # hide
 
 ![](basic_usage.svg)
 
-As it is exemplified above, it is possible to initialize an estimation using
+## Density Estimation for Conformational Samples
+
+Here, we exemplify the use of `ParallelKDE` to estimate the conformational density of alanine dipeptide via dihedral angles.
+
+We start by downloading and reading the dataset of dihedral angles obtained from molecular dynamics trajectories [^1] [^2].
+
+```@example 2
+using Downloads
+using NPZ
+
+url = "http://ftp.imp.fu-berlin.de/pub/cmb-data/alanine-dipeptide-3x250ns-backbone-dihedrals.npz"
+dest = joinpath(pwd(), "ala_dipeptide_dihderals.npz")
+Downloads.download(url, dest)
+
+ala_npz = npzread(dest)
+ala = vcat(values(ala_npz)...)
+# afterwards one may need to subsample the dataset to obtain uncorrelated samples.
+nothing # hide
+```
+
+This time around we will define the grid that we want to use for the estimation instead of using the default grid:
+
+```@example 2
+using ParallelKDE
+
+phi_range, psi_range = fill(range(-π, π, length=250), 2)
+dihedral_grid = initialize_grid([phi_range, psi_range], device=:cpu) # or device=:cuda
+
+density_estimation = initialize_estimation(
+  ala', grid=dihedral_grid, device=:cpu # or device=:cuda
+)
+estimate_density!(density_estimation, :gradepro)
+estimated_density = get_density(density_estimation)
+nothing # hide
+```
+
+Finally, we can create a contour plot of the estimated density:
+
+```@example 2
+ENV["GKSwstype"]="nul" # hide
+using Plots
+using LaTeXStrings
+
+p = contourf(phi_range, psi_range, estimated_density')
+plot!(p, xlabel=L"$\phi$", ylabel=L"$\psi$", colorbar_title="Estimated Density")
+savefig("ala_dihedrals.svg"); nothing # hide
+```
+
+![](ala_dihedrals.svg)
+
+## GradePro vs Rules of Thumb
+
[…]
