### 2025-12-22 41f9f4a
removed pluto notebook
