### 2025-12-23 #9616 @Clownshift
> [@Clownshift](https://github.com/Clownshift) please include 'Community Guidelines' (required) and more examples in the docs for different situations.

Hi @mantzaris , thanks for the feedback.

We now have 'Community Guidelines' in the docs, but I am afraid, they are a bit hidden. You can find them in the [docs of the dev branch](https://chrissm23.github.io/ParallelKDE.jl/dev/contributing/)
Besides that, we also added more examples as per your suggestion. These can be found [here](https://chrissm23.github.io/ParallelKDE.jl/dev/).

We will collect all changes due to comments from the reviewers, in `dev` before we release a new stable version. That way, we do not generate too many updates for the general registry. I hope this is alright for the review process.
