### 2026-02-09 #9616 @hisacro
@jbytecode

I have been testing the usage mentioned on the documentation pages, I believe the main aspect of the package is CUDA implementation. I have resource to run on GPU, but I'm struggling to  find the benchmarked  sample data set that resulted in Fig. 1,2 in the draft.

If I missed anything, is there any documentation page regarding this?
