### 2026-02-09 #9616 @Clownshift
> [@jbytecode](https://github.com/jbytecode)
>
> I have been testing the usage mentioned on the documentation pages, I believe the main aspect of the package is CUDA implementation. I have resource to run on GPU, but I'm struggling to find the benchmarked sample data set that resulted in Fig. 1,2 in the draft.
>
> If I missed anything, is there any documentation page regarding this?

Hi @hisacro ,
Thanks for the thorough review. You are correct, the fact that you can run `ParallelKDE` on GPU is one of the main aspects. We use the `CUDA.jl` interface to run calculations on the GPU through julia.

Apologies that it is not clear from the current version of the manuscript, but we did not store all the data from the benchmark in repository. The main reason is that it is actually quite a lot of data. Also, the benchmark is done with synthetic stochastic data, which can be recreated with common libraries (in julia, or in python). Therefore, you may replicate the benchmark either directly in julia, or through the python-wrapper of `ParallelKDE`, which is called [ParallelKDEpy](https://github.com/chrissm23/ParallelKDEpy). Note that, if you use julia, the other packages which we use for comparison--which are python packages--cannot straightfowardly be used.

In the documentation, we mention at the [main page](https://chrissm23.github.io/ParallelKDE.jl/stable/#ParallelKDE) how to reproduce a single iteration of the benchmark, but we acknowledge that, in the current version of the manuscript, it is not sufficiently clear how the benchmark was performed. To address your comment, we added a short explanation to the manuscript, explaining the benchmark more extensively. You can find the updated version of the manuscript in the [dev-branch](https://github.com/chrissm23/ParallelKDE.jl/tree/dev), in the [updated pdf](https://github.com/chrissm23/ParallelKDE.jl/actions/runs/21828949639/artifacts/5433393718).

**To check the performance**, you may execute the code snippet below. If you want to use the GPU in the benchmark, uncomment `# device=:cuda`. However, be aware that repeating **the whole benchmark will take many hours**, even on very fast GPUs, especially, if you look at 2D-distributions. Therefore, this code snippet only executes one examplary setting not the whole benchmark printed in the manuscript. Meaning, below, we do not test all estimators and not all sample sizes that we tested for the article.

------
code snippet:

```julia
using ParallelKDE
using Distributions
using BenchmarkTools

n_samples = 10000
n_gridpoints = 300
device=:cpu
# device=:cuda

# -- 1D example --
data = randn(1, n_samples) # 1-dimensional sample of 10000 points
grid = initialize_grid(
  [range(-8, 8, length=n_gridpoints)],
  device=device # or :cuda
)

println("1D runtime for $n_samples samples and $n_gridpoints gridpoints:")

@btime begin
  density_estimation = initialize_estimation(
    $data,
    grid=$grid,
    device=$device, # or cuda
  )
  estimate_density!(
    density_estimation,
    :gradepro, # or :rot
  )

  density_estimated = get_density(density_estimation)
end

# -- 2D example --
distro = MvNormal(fill(0.0, 2), oneunit(Matrix{Float64}(undef, 2, 2)))

data = rand(distro, n_samples)
grid = initialize_grid(
  fill(range(-8, 8, length=n_gridpoints), 2),
  device=device, # or cuda
)

println("2D runtime for $n_samples samples and $n_gridpoints gridpoints:")

@btime begin
  density_estimation = initialize_estimation(
    $data,
    grid=$grid,
    device=$device, # or cuda
  )
  estimate_density!(
    density_estimation,
    :gradepro, # or :rot
  )

  density_estimated = get_density(density_estimation)
end
```
