### 2026-02-10 #9616 @hisacro
> Hi [@hisacro](https://github.com/hisacro) , Thanks for the thorough review. You are correct, the fact that you can run `ParallelKDE` on GPU is one of the main aspects. We use the `CUDA.jl` interface to run calculations on the GPU through julia.

Thanks for the quick response, First I need to clarify my sentiment. From my point of JOSS journal, if there are generated image on the manuscript, the scripts/codes needs to be in public repository or in the package's documentation. This practice reflects the essence of JOSS's commitment and makes it stand apart as an open source journal.

> Apologies that it is not clear from the current version of the manuscript, but we did not store all the data from the benchmark in repository. The main reason is that it is actually quite a lot of data. Also, the benchmark is done with synthetic stochastic data, which can be recreated with common libraries (in julia, or in python).

I totally understand it but I believe this needs to be mentioned explicitly in the manuscript.

> Note that, if you use julia, the other packages which we use for comparison--which are python packages--cannot straightfowardly be used.

I'm using Julia, and where these  comparison scripts are present?

>  You can find the updated version of the manuscript in the [dev-branch](https://github.com/chrissm23/ParallelKDE.jl/tree/dev), in the [updated pdf](https://github.com/chrissm23/ParallelKDE.jl/actions/runs/21828949639/artifacts/5433393718).

Thank you. One additional clarification might be on the hardware, I see you have mentioned Intel Core i7-6700, are you using all 4 cores?

>
> **To check the performance**, you may execute the code snippet below. If you want to use the GPU in the benchmark, uncomment `# device=:cuda`. However, be aware that repeating **the whole benchmark will take many

I'm including my benchmark results of both CPU & GPU
```
# CPU, GradePro
2D runtime for 100 samples and 300 gridpoints:
  161.208 s (282135 allocations: 1.56 GiB)
300×300 Matrix{Float64}:

2D runtime for 1000 samples and 300 gridpoints:
   133.482 s (2644642 allocations: 1.64 GiB)
300×300 Matrix{Float64}:

2D runtime for 10000 samples and 300 gridpoints:
  164.661 s (26438289 allocations: 2.38 GiB)
300×300 Matrix{Float64}:

# CPU, ROT
 2D runtime for 100 samples and 300 gridpoints:
  3.565 ms (2726 allocations: 4.90 MiB)
300×300 Matrix{Float64}:

2D runtime for 1000 samples and 300 gridpoints:
  2.699 ms (34136 allocations: 5.82 MiB)
300×300 Matrix{Float64}:

2D runtime for 10000 samples and 300 gridpoints:
  9.601 ms (340136 allocations: 14.88 MiB)
300×300 Matrix{Float64}:

# GPU, GradePro
2D runtime for 100 samples and 300 gridpoints:
  5.422 s (367376 allocations: 11.39 MiB)
300×300 CUDA.CuArray{Float32, 2, CUDA.DeviceMemory}:

2D runtime for 1000 samples and 300 gridpoints:
  6.405 s (368254 allocations: 11.43 MiB)
300×300 CUDA.CuArray{Float32, 2, CUDA.DeviceMemory}:

2D runtime for 10000 samples and 300 gridpoints:
  5.740 s (368255 allocations: 11.67 MiB)
300×300 CUDA.CuArray{Float32, 2, CUDA.DeviceMemory}:

#  GPU, ROT
2D runtime for 100 samples and 300 gridpoints:
  559.833 μs (1566 allocations: 1.42 MiB)
300×300 CUDA.CuArray{Float32, 2, CUDA.DeviceMemory}:

2D runtime for 1000 samples and 300 gridpoints:
  588.487 μs (1684 allocations: 1.43 MiB)
300×300 CUDA.CuArray{Float32, 2, CUDA.DeviceMemory}:

2D runtime for 10000 samples and 300 gridpoints:
  608.277 μs (1684 allocations: 1.57 MiB)
300×300 CUDA.CuArray{Float32, 2, CUDA.DeviceMemory}:
```

These are  run on AMD EPYC 9745 & Nvidia L40S with the following config
```
   NumNodes=1 NumCPUs=4 NumTasks=4 CPUs/Task=1 ReqB:S:C:T=0:0:*:1
   ReqTRES=cpu=4,mem=8G,node=1,billing=4,gres/gpu=1
```

I can see the trend between GradePro & ROT is similar to your benchmarks. Overall I believe the claims are  valid,  but please try to include the script you used for other packages or atleast point the users to those package documentation.
