+++ paper/paper.md
@@ -42,14 +42,15 @@ Many researchers rely on Python implementations of KDE. Popular examples include
 
 # Performance benchmarks
 
-To demonstrate that our approach—moving density estimation to parallel execution or parallel hardware—produces marked acceleration compared with standard CPU-based libraries, we performed an extensive benchmark of different methods. The results are shown in \autoref{fig:benchmark_samples} and \autoref{fig:benchmark_grids}. Our implementations in `ParallelKDE` are particularly fast for high-resolution grids and/or large datasets.
-Additionally, our novel method, GradePro, achieves high accuracy on diverse distributions. While the CPU version is slower than several 1D baselines and comparable in 2D, its performance is still competitive and the extra computation yields better accuracy. In CUDA, the method's parallelism is exposed, making it faster and produces superior scaling with problem size. A detailed introduction of GradePro along with an extensive investigation of its accuracy is provided in @sustay_martinez_novel_2025.
+To demonstrate that our approach—moving density estimation to parallel execution or parallel hardware—produces marked acceleration compared with standard CPU-based libraries, we performed an extensive benchmark in Python, since this is where the most widely used KDE tools are provided. To ensure consistent interface across methods, we ran all experiments through our lightweight wrapper `ParallelKDEpy` (see benchmark scripts in the repository: <https://github.com/chrissm23/ParallelKDE.jl/tree/main/paper>). We evaluated different methods on synthetic data drawn from 1D and 2D standard normal distributions, generated with SciPy. For compatibility with the existing Python tools, we executed CPU baselines single-threaded; in addition, we benchmarked our CUDA implementation to exemplify the acceleration achievable through parallel execution.
+
+The results are shown in \autoref{fig:benchmark_samples} and \autoref{fig:benchmark_grids}. Our implementations in `ParallelKDE` are particularly fast for high-resolution grids and/or large datasets. Additionally, our novel method, GradePro, achieves high accuracy on diverse distributions. While the CPU version is slower than several 1D baselines and comparable in 2D, its performance is still competitive and the extra computation yields better accuracy. In CUDA, the method's parallelism is exposed, making it faster and produces superior scaling with problem size. A detailed introduction of GradePro along with an extensive investigation of its accuracy is provided in @sustay_martinez_novel_2025.
 
 Further comparison of GradePro with other methods is in progress. In the future, we aim to provide general recommendations for users on the optimal choice of bandwidth selection methods for different use-cases. Here, we focus on the software engineering aspects, which enable our fast implementation.
 
-![Benchmark of common KDE packages and their estimators—rule-of-thumb (ROT), plug-in, and GradePro—along with estimators in ParallelKDE at different sample sizes. The estimations were performed on normally distributed data. In 1D, 100 and 100,000 samples were used with a grid of 500 points. In 2D, 1,000 and 1,000,000 samples were used with a grid of 100 points per dimension. Reported runtimes are averages over 10 repetitions. Hardware: Intel Core i7-6700 and NVIDIA GTX 1080.\label{fig:benchmark_samples}](./benchmark_samples.pdf)
[…]
