+++ paper/paper.md
@@ -42,7 +42,7 @@ Many researchers rely on Python implementations of KDE. Popular examples include
 
 # Performance benchmarks
 
-To demonstrate that our approach—moving density estimation to parallel execution or parallel hardware—produces marked acceleration compared with standard CPU-based libraries, we performed an extensive benchmark in Python, since this is where the most widely used KDE tools are provided. To ensure consistent interface across methods, we ran all experiments through our lightweight wrapper `ParallelKDEpy` (see benchmark scripts in the repository: <https://github.com/chrissm23/ParallelKDE.jl/tree/main/paper>). We evaluated different methods on synthetic data drawn from 1D and 2D standard normal distributions, generated with SciPy. For compatibility with the existing Python tools, we executed CPU baselines single-threaded; in addition, we benchmarked our CUDA implementation to exemplify the acceleration achievable through parallel execution.
+To demonstrate that our approach—moving density estimation to parallel execution or parallel hardware—produces marked acceleration compared with standard CPU-based libraries, we performed an extensive benchmark. We chose to carried it out in Python, since the most popular KDE tools are provided for that ecosystem. To ensure consistent interface across methods, we ran all experiments through our lightweight wrapper `ParallelKDEpy`. Naturally, our tool can conveniently be used directly in Julia, too. We evaluated different methods on synthetic data drawn from 1D and 2D standard normal distributions, generated with SciPy. For compatibility with the existing Python tools, we executed CPU baselines single-threaded; in addition, we benchmarked our CUDA implementation to exemplify the acceleration achievable through parallel execution. To reproduce the benchmark, please refer to the corresponding section in the `README.md`.
 
 The results are shown in \autoref{fig:benchmark_samples} and \autoref{fig:benchmark_grids}. Our implementations in `ParallelKDE` are particularly fast for high-resolution grids and/or large datasets. Additionally, our novel method, GradePro, achieves high accuracy on diverse distributions. While the CPU version is slower than several 1D baselines and comparable in 2D, its performance is still competitive and the extra computation yields better accuracy. In CUDA, the method's parallelism is exposed, making it faster and produces superior scaling with problem size. A detailed introduction of GradePro along with an extensive investigation of its accuracy is provided in @sustay_martinez_novel_2025.
 
