### 2026-02-10 #9616 @Clownshift
> Thanks for the quick response, First I need to clarify my sentiment. From my point of JOSS journal, if there are generated image on the manuscript, the scripts/codes needs to be in public repository or in the package's documentation. This practice reflects the essence of JOSS's commitment and makes it stand apart as an open source journal.
>
>
> I totally understand it but I believe this needs to be mentioned explicitly in the manuscript.
> >
> I'm using Julia, and where these comparison scripts are present?
>
>
> Thank you. One additional clarification might be on the hardware, I see you have mentioned Intel Core i7-6700, are you using all 4 cores?
>
> I'm including my benchmark results of both CPU & GPU
>
> ``` ...
> ```
>
> I can see the trend between GradePro & ROT is similar to your benchmarks. Overall I believe the claims are valid, but please try to include the script you used for other packages or atleast point the users to those package documentation.

Hello @hisacro,

Thank you for the clarifications. We also agree that making public any code used to generate images for the manuscript is important to guarantee reproducibility. We are sorry that we did not properly explain earlier where to find the code for the benchmark. The benchmark is fully reproducible from the repository via a [Jupyter notebook](https://github.com/chrissm23/ParallelKDE.jl/blob/main/paper/benchmark_time.ipynb) that runs the complete workflow end-to-end (data generation, method calls and figure generation). Apologies, we should have mentioned this jupyter notebook earlier in the discussion. This notebook is placed in the [`paper/`](https://github.com/chrissm23/ParallelKDE.jl/tree/dev/paper) directory of the repo. There you may also find the corresponding figures, as shown in the paper. To make this notebook easier to find, we added an additional section to the [`README.md`](https://github.com/chrissm23/ParallelKDE.jl/tree/dev?tab=readme-ov-file#benchmarks) of the repository, which we now also reference in the manuscript.

Note that this jupyter notebook is written in Python. The reason we made the comparison in Python instead of Julia is that all the other tools we compare against---which are, to the best of our knowledge, the most popular for KDE calculations---are primarily available in the Python ecosystem. To keep the benchmark consistent across methods, for the benchmark we execute our `ParallelKDE` algorithms using our lightweight wrapper `ParallelKDEpy`. We agree that our explanation  of the benchmark was generally rather short before. Therefore, we extended the explanation in the manuscript and `README.md`, respectively.

Lastly, regarding the hardware, we would like to thank you for pointing out the missing information. We executed the CPU baseline of the benchmark single-threaded for comparability with the other tools. As you suggested, we included that explanation to the manuscript.

For your convenience, please find the updated version of the manuscript [here](https://github.com/chrissm23/ParallelKDE.jl/actions/runs/21865834304/artifacts/5448484009)
