### 2026-02-10 #9616 @Clownshift
> [@editorialbot](https://github.com/editorialbot) generate pdf
>
> Please use editorial bot to generate new versions of the manuscript and avoid attaching external files or links. [@Clownshift](https://github.com/Clownshift)

Dear @jbytecode ,
Apologies, I will stop linking the new  versions of the paper.

We will merge all the latest changes from `dev` to `main`. The preprint generated by the "editorialbot", will be up to date then.
