### 2026-02-19 c9d8269
minor fixes in bibtex and manuscript
