+++ paper/paper.bib
@@ -55,9 +55,9 @@
 }
 
 @misc{statsmodels_developers_statsmodelsnonparametrickernel_densitykdemultivariate_2024,
-	title = {statsmodels.nonparametric.kernel\_density.{KDEMultivariate} - statsmodels v0.14.4 {Documentation}},
+	title = {statsmodels v0.14.4 {Documentation}},
 	url = {https://www.statsmodels.org/stable/generated/statsmodels.nonparametric.kernel_density.KDEMultivariate.html},
-	author = {{statsmodels developers}},
+	author = {{Statsmodels Developers}},
 	year = {2024},
 }
 
