### 2026-02-19 6830f34
minor fixes in bibtex and manuscript
