### 2026-02-19 #9616 @jbytecode
@Clownshift - I've just sent a pull request:

https://github.com/chrissm23/ParallelKDE.jl/pull/27

This PR includes minor fixes and suggestions in both the bibliography and the manuscript. Please review the changes. If you're agree with them, please merge.

The other issue is the size of Figure 3. (which has the caption of Flowchart of the parallelizable point-wise density estimation algorithm). Is it possible to reduce the height as it perfectly fit the page?

Thank you in advance.
