### 2026-02-19 8cebd05
decreased fig height
