### 2026-02-19 #9616 @jbytecode
@Clownshift - The next steps are as follows:

- Please create a tagged release in the repo. This tagged release will be a proof of the reviewed software.
- Use this tagged release to create a Zenodo archive.
- In this Zenodo archive, the title, author list and their metadata should match those in the paper.
- Please report me the tagged release info, Zenodo page, and the DOI here

Thank you in advance.
