### 2026-02-20 #9616 @Clownshift
> [@Clownshift](https://github.com/Clownshift) - The next steps are as follows:
>
> * Please create a tagged release in the repo. This tagged release will be a proof of the reviewed software.
> * Use this tagged release to create a Zenodo archive.
> * In this Zenodo archive, the title, author list and their metadata should match those in the paper.
> * Please report me the tagged release info, Zenodo page, and the DOI here
>
> Thank you in advance.

@jbytecode , thank you for handling the editorial work so swiftly and smoothly. It's a pleasure working with you!

We actually found a small typo, which has been corrected in this [pull request](https://github.com/chrissm23/ParallelKDE.jl/pull/28).

As requested,
- [the tagged release](https://github.com/chrissm23/ParallelKDE.jl/releases/tag/v1.0.2)
- [the Zenodo archive](https://zenodo.org/records/18712217)
-  the doi of the Zenodo archive is: [10.5281/zenodo.18712216](https://doi.org/10.5281/zenodo.18712216)
