### 2026-02-20 #9616 @jbytecode
@Clownshift - The title of the paper is

"ParallelKDE.jl: A Package for Highly Parallel Kernel Density Estimation"

whereas the title of the archive is

"ParallelKDE.jl: A Package for Highly Parallel Kernel Density Estimation (v1.0.2)"

Please remove the version info.

The license of the software is MIT in the repo, but it's Creative Commons 4 in the Zenodo archive. Please fix them and then fix me again.

Thank you in advance.
