+++ paper/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 Kernel density estimation (KDE) is a valuable tool in exploratory analysis, simulation, and probabilistic modeling across the sciences. However, its susceptibility to the curse of dimensionality can make routine KDE prohibitively slow on large, high-dimensional datasets. A central problem in KDE is the choice of the bandwidth of the kernels, which is oftentimes solved by applying an empiric rule-of-thumb (ROT). Furthermore, estimation of non-normal distributions is sometimes approached through non-constant bandwidths—so called adaptive bandwith—which adds an additional layer of complication. For additional background on KDE and bandwidth selection, see @silverman_density_1986, @scott_multivariate_1992, @wand_kernel_1995 and @jones_brief_1996.
 
-`ParallelKDE.jl` provides a fast, open-source KDE toolkit in Julia [@bezanson_julia_2017] that provides extensible infrastructure to support a variety of methods with serial, multi-threaded and GPU implementations. The package includes (i) a highly parallel implementation of established KDE rules-of-thumb (ROTs) as well as (ii) an implementation of GradePro, a new algorithm described in [@sustay_martinez_novel_2025], engineered here for serial, threaded, and CUDA backends. A lightweight Python wrapper, `ParallelKDEpy` exposes the same functionality to Python workflows. Our goal is to make robust density estimation practical for large sample sizes of moderate dimensionality through careful parallelization of, mainly, grid-based KDE estimators.
+`ParallelKDE.jl` provides a fast, open-source KDE toolkit in Julia [@bezanson_julia_2017] that provides extensible infrastructure to support a variety of methods with serial, multi-threaded and GPU implementations. The package includes (i) a highly parallel implementation of established KDE rules-of-thumb (ROTs) as well as (ii) an implementation of GradePro, a new algorithm described in @sustay_martinez_novel_2025, engineered here for serial, threaded, and CUDA backends. A lightweight Python wrapper, `ParallelKDEpy` exposes the same functionality to Python workflows. Our goal is to make robust density estimation practical for large sample sizes of moderate dimensionality through careful parallelization of, mainly, grid-based KDE estimators.
 
 # Statement of need
 
