### 2026-03-06 #9616 @Clownshift
Hi @crvernon ,

Thank you for your nice feedback.

We corrected the formatting of this citation, as you suggested. This change has been pushed to `main` and `dev`.

Thank you for this smooth reviewing process!
